Revert "Use fs trait, but going to revert this because no point."
This reverts commit 3be7ba9131d9cc69ea0018b75b52d482ed1fc6e8.
