+++ cli/cache/http_cache/common.rs
@@ -1,5 +1,7 @@
 // Copyright 2018-2023 the Deno authors. All rights reserved. MIT license.
 
+use std::path::Path;
+
 use deno_core::url::Url;
 
 pub fn base_url_to_filename_parts(
@@ -29,4 +31,27 @@ pub fn base_url_to_filename_parts(
   };
 
   Some(out)
+}
+
+pub fn read_file_bytes(path: &Path) -> std::io::Result<Option<Vec<u8>>> {
+  match std::fs::read(path) {
+    Ok(s) => Ok(Some(s)),
+    Err(err) if err.kind() == std::io::ErrorKind::NotFound => Ok(None),
+    Err(err) => Err(err),
+  }
+}
+
+/// Ensures the location of the cache.
+pub fn ensure_dir_exists(path: &Path) -> std::io::Result<()> {
+  if path.is_dir() {
+    return Ok(());
+  }
+  std::fs::create_dir_all(path).map_err(|e| {
+    std::io::Error::new(
+      e.kind(),
+      format!(
+        "Could not create remote modules cache location: {path:?}\nCheck the permission of the directory."
+      ),
+    )
+  })
 }
