+++ cli/factory.rs
@@ -250,8 +250,7 @@ impl CliFactory {
       let global_cache = self.global_http_cache()?.clone();
       match self.options.remote_modules_dir_path() {
         Some(local_path) => {
-          let local_cache =
-            LocalHttpCache::new(local_path, global_cache, self.fs().clone());
+          let local_cache = LocalHttpCache::new(local_path, global_cache);
           Ok(Arc::new(local_cache))
         }
         None => Ok(global_cache),
